on page unload try to save pending state
I realized this by using `onunload` on `window` and making an ajax call with the pending state.
Note that I used busy waiting at the end because the browser would kill any asynchronous call.
This obviously causes some delay when closing the browser on pending changes.
I changed the debounce to 5 seconds to submit changes to MAPPING and the same maximal delay at browser close/reload.

closes T69