Added single run to the headless tests, such that the docker build terminates when all tests were finished successfully.
Dockerfile now runs the tests. Added g++ and make for npm install (failed on my system without g++), xvfb as virtual framebuffer, python for npm install and some other browsers (TODO) and firefox to the imports.
We then run Xvfb as background task and tell the tests to take the given DISPLAY environment variable on which xvfb simulates a display.